+++ docx4j-samples-docx-export-fo/src/main/java/org/docx4j/samples/ManyFactory.java
@@ -16,6 +16,7 @@ import org.apache.fop.apps.FopFactoryBuilder;
 import org.docx4j.Docx4jProperties;
 import org.docx4j.convert.out.FOSettings;
 import org.docx4j.convert.out.common.Exporter;
+import org.docx4j.convert.out.fo.FOExporterVisitor;
 import org.docx4j.convert.out.fo.FOExporterXslt;
 import org.docx4j.convert.out.fo.renderers.FORendererApacheFOP;
 import org.docx4j.fonts.IdentityPlusMapper;
@@ -129,8 +130,8 @@ public class ManyFactory {
         try (OutputStream os = new FileOutputStream(outputfilepath)) {
         	
             // Instantiate your chosen exporter manually
-            Exporter<FOSettings> exporter = FOExporterXslt.getInstance();     // XSLT: Fully featured
-//          Exporter<FOSettings> exporter = FOExporterVisitor.getInstance();  // Non XSLT: Faster, but fewer features
+//          Exporter<FOSettings> exporter = FOExporterXslt.getInstance();     // XSLT: Fully featured, old default
+            Exporter<FOSettings> exporter = FOExporterVisitor.getInstance();  // Non XSLT: Faster, and fully featured since 17.0.4
             
             // Execute programmatic conversion directly to the target outputstream
             exporter.export(foSettings, os);